phase3: split lookup table functions into src/lookup_tables.cpp
